This is part of #1674, currently `benchcmp` looks like this:
I might drop the `TestCase` usage for benchmarks as it might affect the results, the dispatcher seems to perform a bit faster (are the allocations made by the test case code? still investigating this):
```
% benchcmp old new
benchmark                                            old ns/op     new ns/op     delta
BenchmarkGRPCDispatch/Pre_Hook_with_SetHeaders-8     558806        526713        -5.74%

benchmark                                            old allocs     new allocs     delta
BenchmarkGRPCDispatch/Pre_Hook_with_SetHeaders-8     669            803            +20.03%

benchmark                                            old bytes     new bytes     delta
BenchmarkGRPCDispatch/Pre_Hook_with_SetHeaders-8     84621         89969         +6.32%
```
I will add more scenarios and also an API with multiple hooks.