@matiasinsaurralde 2018-05-23
Have removed `TestCase` (now using `http.Get`/`http.Post` directly)  from the benchmarks:
```
goos: darwin
goarch: amd64
pkg: github.com/TykTechnologies/tyk
BenchmarkGRPCDispatch/Pre_Hook_with_SetHeaders-8         	   50000	     20316 ns/op	    2474 B/op	      43 allocs/op
```