@buger 2018-05-30
Allocations are quite small anyway. Looks good.